add AbstractPage
