+++ src/test/java/cases/AbstractTest.java
@@ -29,5 +29,4 @@ public class AbstractTest {
         wait.until(ExpectedConditions.elementToBeClickable(By.xpath(loginPage.loginButton)));
         loginPage.loginAsAdmin();
     }
-
 }
