+++ src/test/java/pages/AdminGeoZonesPage.java
@@ -1,17 +1,23 @@
 package pages;
 
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.util.List;
 
-public class AdminGeoZonesPage extends AdminCountriesPage {
+public class AdminGeoZonesPage extends AbstractPage {
 
     //Locators
     private final String geoZoneNamesElement = "//td/a[@href][not(@title=\"Edit\")]";
+    private final String geoZonePageElement = "//form[@name=\"geo_zones_form\"]";
 
     public AdminGeoZonesPage(WebDriver driver) {
-        super(driver);
+        this.driver = driver;
+        wait = new WebDriverWait(driver, 10);
+        wait.until(ExpectedConditions.presenceOfElementLocated(By.xpath(geoZonePageElement)));
     }
 
     public List<WebElement> getGeoZonesList() {
